feat: libraryInfo 엔티티를 생성한다. (#13)
