+++ core-module/src/main/java/com/example/coremodule/book/BookRepository.java
@@ -1,4 +1,4 @@
-package com.example.coremodule.book.domain;
+package com.example.coremodule.book;
 
 import org.springframework.data.jpa.repository.JpaRepository;
