+++ core-module/src/main/java/com/example/coremodule/library_info/LibraryInfoRepository.java
@@ -1,0 +1,6 @@
+package com.example.coremodule.library_info;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LibraryInfoRepository extends JpaRepository<LibraryInfo, Long> {
+}
